requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.20.0
pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
seaborn>=0.11.0
streamlit>=1.20.0
